refactored files, added templates
