+++ .gitignore
@@ -1,4 +1,14 @@
-__pycache__/
-*pyc
-*egg-info/
 .venv/
+
+*.pyc
+__pycache__/
+
+instance/
+
+.pytest_cache/
+.coverage
+htmlcov/
+
+dist/
+build/
+*.egg-info/
